+++ OpenTabletDriver.Native/Linux/Timers/SigEv.cs
@@ -1,0 +1,10 @@
+namespace OpenTabletDriver.Native.Linux.Timers
+{
+    public enum SigEv
+    {
+        Signal,
+        None,
+        Thread,
+        ThreadID
+    }
+}
