+++ OpenTabletDriver.Native/Linux/Timers/Structs/SigVal.cs
@@ -1,0 +1,15 @@
+using System;
+using System.Runtime.InteropServices;
+
+namespace OpenTabletDriver.Native.Linux.Timers.Structs
+{
+    [StructLayout(LayoutKind.Explicit)]
+    public struct SigVal
+    {
+        [FieldOffset(0)]
+        public int sival_int;
+
+        [FieldOffset(0)]
+        public IntPtr sival_ptr;
+    }
+}
