+++ OpenTabletDriver.Native/Linux/Timers/TimerFlag.cs
@@ -1,0 +1,8 @@
+namespace OpenTabletDriver.Native.Linux.Timers
+{
+    public enum TimerFlag
+    {
+        Default,
+        AbsoluteTime
+    }
+}
