+++ OpenTabletDriver.Plugin/ITimer.cs
@@ -6,7 +6,6 @@ namespace OpenTabletDriver.Plugin.Timers
     {
         public void Start();
         public void Stop();
-        public bool Stop(int milliseconds);
         public bool Enabled { get; }
         public float Interval { get; set; }
         public event Action Elapsed;
